Move goal-general functions to GoalActions, merge-specific functions to MergeDupStepActions.
